Quality of Life Updates: Better JS performance, I have split up the script.js file into individual files based on their function. More important scripts will be loaded first, while less important will be last or even defered
Better video support, to eliminate the need of Adobe Premiere Pro or Adobe After Effects, I had changed the CSS to place the iMac frame over in the browser. This eliminates the need for it to be added in post.
This change will allow me to update content more rapidly, it will remove a tedious step and allow me to on the fly fix typo's and update content.

More changes will be coming soon to make the site more enjoyable

Soon ill finish the rest of the videos (next commit) and make the canonical URL based on the path.